+++ src/spatialdata/_core/spatialdata.py
@@ -628,7 +628,7 @@ class SpatialData:
         found_element_name: list[str] = []
         for element_type in ["images", "labels", "points", "shapes", "tables"]:
             for element_name, element_value in getattr(self, element_type).items():
-                if id(element_value) == id(element):
+                if element_value is element:
                     found.append(element_value)
                     found_element_type.append(element_type)
                     found_element_name.append(element_name)
@@ -2225,6 +2225,7 @@ class QueryManager:
         polygon: Polygon | MultiPolygon,
         target_coordinate_system: str,
         filter_table: bool = True,
+        clip: bool = False,
     ) -> SpatialData:
         """
         Perform a polygon query on the SpatialData object.
@@ -2239,6 +2240,7 @@ class QueryManager:
             polygon=polygon,
             target_coordinate_system=target_coordinate_system,
             filter_table=filter_table,
+            clip=clip,
         )
 
     def __call__(self, request: BaseSpatialRequest, **kwargs) -> SpatialData:  # type: ignore[no-untyped-def]