+++ tests/core/operations/test_aggregations.py
@@ -8,6 +8,7 @@ from anndata import AnnData
 from anndata.tests.helpers import assert_equal
 from geopandas import GeoDataFrame
 from numpy.random import default_rng
+
 from spatialdata import aggregate, to_polygons
 from spatialdata._core._deepcopy import deepcopy as _deepcopy
 from spatialdata._core.spatialdata import SpatialData
@@ -366,7 +367,7 @@ def test_aggregate_requiring_alignment(sdata_blobs: SpatialData, values, by) -> 
         raise pytest.skip("Aggregation mixing raster and vector data is not currently supported.")
     values = sdata_blobs[values]
     by = sdata_blobs[by]
-    if id(values) == id(by):
+    if values is by:
         # warning: this will give problems when aggregation labels by labels (not supported yet), because of this: https://github.com/scverse/spatialdata/issues/269
         by = _deepcopy(by)
         assert by.attrs["transform"] is not values.attrs["transform"]