+++ tests/core/operations/test_map.py
@@ -6,7 +6,6 @@ import numpy as np
 import pytest
 from spatialdata._core.operations.map import _relabel_sequential, map_raster
 from spatialdata.transformations import Translation, get_transformation, set_transformation
-from xarray import DataArray
 
 
 def _multiply(arr, parameter=10):